# stt-service/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "STT Service API",
    "description": "Whisper-powered speech-to-text transcription service with intelligent segmentation and multi-language support",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "STT Service (External Port)"
    }
  ],
  "paths": {
    "/transcribe": {
      "post": {
        "summary": "Transcribe audio to text",
        "description": "Convert audio file to text using Whisper with optional segmentation for training data preparation",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file to transcribe"
                  },
                  "language": {
                    "type": "string",
                    "description": "Target language code (auto-detect if not specified)",
                    "example": "en"
                  },
                  "model": {
                    "type": "string",
                    "enum": ["tiny", "base", "small", "medium", "large"],
                    "default": "medium",
                    "description": "Whisper model size"
                  },
                  "segment_length": {
                    "type": "integer",
                    "description": "Segment length in seconds for training data",
                    "default": 10
                  },
                  "overlap": {
                    "type": "number",
                    "description": "Overlap between segments in seconds",
                    "default": 1.0
                  }
                },
                "required": ["audio_file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful transcription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/detect_language": {
      "post": {
        "summary": "Detect audio language",
        "description": "Automatically detect the language of provided audio file",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file for language detection"
                  }
                },
                "required": ["audio_file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Language detection result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageDetectionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TranscriptionResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Complete transcribed text"
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranscriptionSegment"
            },
            "description": "Segmented transcription with timestamps"
          },
          "language": {
            "type": "string",
            "description": "Detected or specified language"
          },
          "duration": {
            "type": "number",
            "description": "Audio duration in seconds"
          },
          "processing_time": {
            "type": "number",
            "description": "Processing time in seconds"
          }
        }
      },
      "TranscriptionSegment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Segment identifier"
          },
          "start": {
            "type": "number",
            "description": "Start time in seconds"
          },
          "end": {
            "type": "number",
            "description": "End time in seconds"
          },
          "text": {
            "type": "string",
            "description": "Segment text"
          },
          "confidence": {
            "type": "number",
            "description": "Transcription confidence score"
          }
        }
      },
      "LanguageDetectionResponse": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "description": "Detected language code"
          },
          "confidence": {
            "type": "number",
            "description": "Detection confidence score"
          },
          "probabilities": {
            "type": "object",
            "description": "Language probabilities",
            "additionalProperties": {
              "type": "number"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Detailed error information"
          }
        }
      }
    }
  }
}